+++ supvisors/tests/test_viewhandler.py
@@ -259,41 +259,49 @@ def test_write_nav_instances_silent_instance(handler):
 
 def test_write_nav_instances_running_instance(handler):
     """ Test the write_nav_instances method using a RUNNING instance. """
-    # set context
-    status = handler.sup_ctx.instances['10.0.0.1']
-    status._state = SupvisorsInstanceStates.RUNNING
     # patch the meld elements
     instance_a_mid = create_element()
     instance_elt = create_element({'instance_a_mid': instance_a_mid})
     instance_li_mid = create_element()
     instance_li_mid.repeat.return_value = [(instance_elt, '10.0.0.1')]
     mocked_root = create_element({'instance_li_mid': instance_li_mid})
-    # test call with address status set in context, RUNNING, different from parameter and not MASTER
     handler.view_ctx = Mock(**{'format_url.return_value': 'an url'})
-    handler.write_nav_instances(mocked_root, '10.0.0.2')
-    assert mocked_root.findmeld.call_args_list == [call('instance_li_mid')]
-    assert instance_li_mid.repeat.call_args_list == [call(list(handler.supvisors.supvisors_mapper.instances.keys()))]
-    assert instance_elt.attrib['class'] == 'RUNNING'
-    assert instance_elt.findmeld.call_args_list == [call('instance_a_mid')]
-    assert handler.view_ctx.format_url.call_args_list == [call('10.0.0.1', 'proc_instance.html')]
-    assert instance_a_mid.attributes.call_args_list == [call(href='an url')]
-    assert instance_a_mid.attrib['class'] == 'on'
-    assert instance_a_mid.content.call_args_list == [call('10.0.0.1')]
-    instance_elt.reset_all()
-    mocked_root.reset_all()
-    handler.view_ctx.format_url.reset_mock()
-    # test call with address status set in context, RUNNING, identical to parameter and MASTER
-    status.state_modes.starting_jobs = True
-    handler.sup_ctx.master_identifier = '10.0.0.1'
-    handler.write_nav_instances(mocked_root, '10.0.0.1')
-    assert mocked_root.findmeld.call_args_list == [call('instance_li_mid')]
-    assert instance_li_mid.repeat.call_args_list == [call(list(handler.supvisors.supvisors_mapper.instances.keys()))]
-    assert instance_elt.attrib['class'] == 'RUNNING active'
-    assert instance_elt.findmeld.call_args_list == [call('instance_a_mid')]
-    assert handler.view_ctx.format_url.call_args_list == [call('10.0.0.1', 'proc_instance.html')]
-    assert instance_a_mid.attributes.call_args_list == [call(href='an url')]
-    assert instance_a_mid.attrib['class'] == 'blink on'
-    assert instance_a_mid.content.call_args_list == [call(f'{MASTER_SYMBOL} 10.0.0.1')]
+    # loop on active states
+    status = handler.sup_ctx.instances['10.0.0.1']
+    all_identifiers = list(handler.supvisors.supvisors_mapper.instances.keys())
+    for state in [SupvisorsInstanceStates.CHECKING, SupvisorsInstanceStates.CHECKED, SupvisorsInstanceStates.RUNNING]:
+        # set context
+        status._state = state
+        status.state_modes.starting_jobs = False
+        handler.sup_ctx.master_identifier = ''
+        # test call with address status set in context, RUNNING, different from parameter and not MASTER
+        handler.write_nav_instances(mocked_root, '10.0.0.2')
+        assert mocked_root.findmeld.call_args_list == [call('instance_li_mid')]
+        assert instance_li_mid.repeat.call_args_list == [call(all_identifiers)]
+        assert instance_elt.attrib['class'] == state.name
+        assert instance_elt.findmeld.call_args_list == [call('instance_a_mid')]
+        assert handler.view_ctx.format_url.call_args_list == [call('10.0.0.1', 'proc_instance.html')]
+        assert instance_a_mid.attributes.call_args_list == [call(href='an url')]
+        assert instance_a_mid.attrib['class'] == 'on'
+        assert instance_a_mid.content.call_args_list == [call('10.0.0.1')]
+        instance_elt.reset_all()
+        mocked_root.reset_all()
+        handler.view_ctx.format_url.reset_mock()
+        # test call with address status set in context, RUNNING, identical to parameter and MASTER
+        status.state_modes.starting_jobs = True
+        handler.sup_ctx.master_identifier = '10.0.0.1'
+        handler.write_nav_instances(mocked_root, '10.0.0.1')
+        assert mocked_root.findmeld.call_args_list == [call('instance_li_mid')]
+        assert instance_li_mid.repeat.call_args_list == [call(all_identifiers)]
+        assert instance_elt.attrib['class'] == state.name + ' active'
+        assert instance_elt.findmeld.call_args_list == [call('instance_a_mid')]
+        assert handler.view_ctx.format_url.call_args_list == [call('10.0.0.1', 'proc_instance.html')]
+        assert instance_a_mid.attributes.call_args_list == [call(href='an url')]
+        assert instance_a_mid.attrib['class'] == 'blink on'
+        assert instance_a_mid.content.call_args_list == [call(f'{MASTER_SYMBOL} 10.0.0.1')]
+        instance_elt.reset_all()
+        mocked_root.reset_all()
+        handler.view_ctx.format_url.reset_mock()
 
 
 def test_write_nav_applications_initialization(handler):